Move records to pagination folder
