+++ LMS.Blazor/LMS.Blazor.Client/Components/TeacherDashboard/TeacherDashboard.razor
@@ -1,4 +1,4 @@
-﻿@using LMS.Shared
+﻿@using LMS.Shared.Pagination
 @inject IApiService ApiService
 @inject NavigationManager NavigationManager
 @attribute [Authorize]
