+++ LMS.Blazor/LMS.Blazor.Client/Pages/DemoAuth.razor
@@ -1,5 +1,5 @@
 ﻿@page "/demoauth"
-@using LMS.Shared
+@using LMS.Shared.Pagination
 @inject IApiService ApiService
 @* @rendermode InteractiveWebAssembly *@                                               @* true *@
 @* @rendermode @(new InteractiveWebAssemblyRenderMode)  *@                             @* true *@
