+++ LMS.Blazor/LMS.Blazor.Client/Pages/Modules.razor
@@ -1,5 +1,5 @@
 @page "/Modules"
-@using LMS.Shared
+@using LMS.Shared.Pagination
 @attribute [Authorize(Roles = "Teacher")]
 @rendermode @(new InteractiveWebAssemblyRenderMode(prerender: false))
 @inject IApiService ApiService
