+++ LMS.Blazor/LMS.Blazor.Client/Pages/Users.razor
@@ -1,5 +1,5 @@
 ﻿@page "/Users"
-@using LMS.Shared
+@using LMS.Shared.Pagination
 @inject IApiService ApiService
 @attribute [Authorize(Roles = RolesNames.Teacher)]
