+++ LMS.Presentation/Controllers/CoursesController.cs
@@ -1,4 +1,5 @@
-﻿using LMS.Shared.Constants;
+﻿using System.Security.Claims;
+using LMS.Shared.Constants;
 using LMS.Shared.DTOs.CourseDtos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -6,8 +7,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Service.Contracts;
 using Swashbuckle.AspNetCore.Annotations;
-using System.Security.Claims;
-using LMS.Shared;
 
 namespace LMS.Presentation.Controllers;
