+++ LMS.Services/ActivityService.cs
@@ -2,8 +2,8 @@ using AutoMapper;
 using Domain.Contracts.Repositories;
 using Domain.Models.Entities;
 using Domain.Models.Exceptions;
-using LMS.Shared;
 using LMS.Shared.DTOs.ActivityDtos;
+using LMS.Shared.Pagination;
 using Service.Contracts;
 
 namespace LMS.Services;
