+++ LMS.Shared/DTOs/CourseDtos/AllCoursesParams.cs
@@ -1,4 +1,6 @@
-﻿namespace LMS.Shared.DTOs.CourseDtos;
+﻿using LMS.Shared.Pagination;
+
+namespace LMS.Shared.DTOs.CourseDtos;
 
 public record AllCoursesParams(
     string? Search,
